refactor(layoutWallet): redux-toolkit migrate done
