+++ packages/blockchain-wallet-v4-frontend/src/data/components/actionTypes.ts
@@ -7,7 +7,6 @@ import * as fundRecovery from './fundRecovery/actionTypes'
 import * as identityVerification from './identityVerification/actionTypes'
 import * as importBtcAddress from './importBtcAddress/actionTypes'
 import * as interest from './interest/actionTypes'
-import * as layoutWallet from './layoutWallet/actionTypes'
 import * as lockbox from './lockbox/actionTypes'
 import * as manageAddresses from './manageAddresses/actionTypes'
 import * as onboarding from './onboarding/actionTypes'
@@ -41,7 +40,6 @@ export {
   identityVerification,
   importBtcAddress,
   interest,
-  layoutWallet,
   lockbox,
   manageAddresses,
   onboarding,
