+++ packages/blockchain-wallet-v4-frontend/src/data/components/actions.ts
@@ -8,7 +8,7 @@ import * as fundRecovery from './fundRecovery/actions'
 import * as identityVerification from './identityVerification/actions'
 import * as importBtcAddress from './importBtcAddress/actions'
 import * as interest from './interest/actions'
-import * as layoutWallet from './layoutWallet/actions'
+import { actions as layoutWallet } from './layoutWallet/slice'
 import * as lockbox from './lockbox/actions'
 import * as manageAddresses from './manageAddresses/actions'
 import * as onboarding from './onboarding/actions'
