--- packages/blockchain-wallet-v4-frontend/src/data/components/layoutWallet/actionTypes.js
@@ -1,7 +1,0 @@
-export const LAYOUT_WALLET_MENU_TOGGLE_CLICKED =
-  '@EVENT.LAYOUT_WALLET_MENU_TOGGLED_CLICKED'
-
-export const LAYOUT_WALLET_MENU_CLOSE_CLICKED =
-  '@EVENT.LAYOUT_WALLET_MENU_CLOSE_CLICKED'
-
-export const SET_BALANCES_CHART_TAB = '@EVENT.SET_BALANCES_CHART_TAB'
