--- packages/blockchain-wallet-v4-frontend/src/data/components/layoutWallet/actions.js
@@ -1,14 +1,0 @@
-import * as AT from './actionTypes'
-
-export const layoutWalletMenuToggleClicked = () => ({
-  type: AT.LAYOUT_WALLET_MENU_TOGGLE_CLICKED
-})
-
-export const layoutWalletMenuCloseClicked = () => ({
-  type: AT.LAYOUT_WALLET_MENU_CLOSE_CLICKED
-})
-
-export const setBalancesTableTab = payload => ({
-  type: AT.SET_BALANCES_CHART_TAB,
-  payload
-})
