+++ packages/blockchain-wallet-v4-frontend/src/data/components/reducers.ts
@@ -4,7 +4,7 @@ import { brokerageReducer as brokerage } from './brokerage/slice'
 import { fundRecoveryReducer } from './fundRecovery/reducers'
 import identityVerificationReducer from './identityVerification/reducers'
 import interestReducer from './interest/reducers'
-import layoutWallet from './layoutWallet/reducers'
+import { layoutWalletReducer as layoutWallet } from './layoutWallet/slice'
 import lockbox from './lockbox/reducers'
 import manageAddresses from './manageAddresses/reducers'
 import { priceChartReducer } from './priceChart/reducers'
